requirements.txt
djangorestframework==3.14.0
gunicorn==21.2.0
python-dotenv==1.0.0
stripe==7.7.0
psycopg2-binary==2.9.5